Add getDocList function to retrieve list of documents for assignment
We create getDocList which retrieves a list of students and documents in order to allow the UI to create a dropdown of student choices.

This function relies on a modified getFirstDocument which now grabs that data at the same time it's opening the doctopus spreadsheet to grab the first document ID. While we're there, we persist data in userProperties 'docData'.