Add dropdown for jumping between students
We add a dropdown menu to allow users to jump between students. The dropdown menu also serves as a handy heading in the UI to tell you which student you're grading.

Changes include:
CSS to style dropdown menu
A callback to grab the list of documents when loading the first time

New function populateDocChooser which populates our dropdown menu.
HTML to insert pick-list at top of commenting UI